# DemoOpenApi/quan/MISA.OpenAPI.service/src/MISA.OAS.Finance.Service/Models/ResponseResult.cs
/*
 * Employees
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;
using MISA.OAS.Finance.Service.Converters;

namespace MISA.OAS.Finance.Service.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class ResponseResult : IEquatable<ResponseResult>
    {
        /// <summary>
        /// Gets or Sets ResultMessage
        /// </summary>
        [DataMember(Name="resultMessage", EmitDefaultValue=false)]
        public string ResultMessage { get; set; }

        /// <summary>
        /// Gets or Sets ResultStream
        /// </summary>
        [DataMember(Name="resultStream", EmitDefaultValue=false)]
        public string ResultStream { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class ResponseResult {\n");
            sb.Append("  ResultMessage: ").Append(ResultMessage).Append("\n");
            sb.Append("  ResultStream: ").Append(ResultStream).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (obj is null) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((ResponseResult)obj);
        }

        /// <summary>
        /// Returns true if ResponseResult instances are equal
        /// </summary>
        /// <param name="other">Instance of ResponseResult to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(ResponseResult other)
        {
            if (other is null) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    ResultMessage == other.ResultMessage ||
                    ResultMessage != null &&
                    ResultMessage.Equals(other.ResultMessage)
                ) && 
                (
                    ResultStream == other.ResultStream ||
                    ResultStream != null &&
                    ResultStream.Equals(other.ResultStream)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (ResultMessage != null)
                    hashCode = hashCode * 59 + ResultMessage.GetHashCode();
                    if (ResultStream != null)
                    hashCode = hashCode * 59 + ResultStream.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(ResponseResult left, ResponseResult right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(ResponseResult left, ResponseResult right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}